refactor: separate core algorithms into modules for clustering models;
