+++ pypots/clustering/vader/model.py
@@ -15,259 +15,16 @@ from typing import Tuple, Union, Optional
 
 import numpy as np
 import torch
-import torch.nn as nn
 from scipy.stats import multivariate_normal
 from sklearn.mixture import GaussianMixture
 from torch.utils.data import DataLoader
 
 from .data import DatasetForVaDER
-from .modules import (
-    inverse_softplus,
-    GMMLayer,
-    PeepholeLSTMCell,
-    ImplicitImputation,
-)
+from .modules import inverse_softplus, _VaDER
 from ..base import BaseNNClusterer
 from ...optim.adam import Adam
 from ...optim.base import Optimizer
 from ...utils.logging import logger
-from ...utils.metrics import cal_mse
-
-
-class _VaDER(nn.Module):
-    """
-
-    Parameters
-    ----------
-    n_steps :
-    d_input :
-    n_clusters :
-    d_rnn_hidden :
-    d_mu_stddev :
-    eps :
-    alpha :
-        Weight of the latent loss.
-        The final loss = `alpha`*latent loss + reconstruction loss
-
-
-    Attributes
-    ----------
-
-    """
-
-    def __init__(
-        self,
-        n_steps: int,
-        d_input: int,
-        n_clusters: int,
-        d_rnn_hidden: int,
-        d_mu_stddev: int,
-        eps: float = 1e-9,
-        alpha: float = 1.0,
-    ):
-        super().__init__()
-        self.n_steps = n_steps
-        self.d_input = d_input
-        self.n_clusters = n_clusters
-        self.d_rnn_hidden = d_rnn_hidden
-        self.d_mu_stddev = d_mu_stddev
-        self.eps = eps
-        self.alpha = alpha
-
-        # building model components
-        self.implicit_imputation_layer = ImplicitImputation(d_input)
-        self.encoder = PeepholeLSTMCell(d_input, d_rnn_hidden)
-        self.decoder = PeepholeLSTMCell(d_input, d_rnn_hidden)
-        self.ae_encode_layers = nn.Sequential(
-            nn.Linear(d_rnn_hidden, d_rnn_hidden), nn.Softplus()
-        )
-        self.ae_decode_layers = nn.Sequential(
-            nn.Linear(d_mu_stddev, d_rnn_hidden), nn.Softplus()
-        )
-        self.mu_layer = nn.Linear(d_rnn_hidden, d_mu_stddev)  # layer for mean
-        self.stddev_layer = nn.Linear(
-            d_rnn_hidden, d_mu_stddev
-        )  # layer for standard variance
-        self.rnn_transform_layer = nn.Linear(d_rnn_hidden, d_input)
-        self.gmm_layer = GMMLayer(d_mu_stddev, n_clusters)
-
-    @staticmethod
-    def z_sampling(
-        mu_tilde: torch.Tensor,
-        stddev_tilde: torch.Tensor,
-    ) -> torch.Tensor:
-        noise = mu_tilde.data.new(mu_tilde.size()).normal_()
-        z = torch.add(mu_tilde, torch.exp(0.5 * stddev_tilde) * noise)
-        return z
-
-    def encode(
-        self,
-        X: torch.Tensor,
-        missing_mask: torch.Tensor,
-    ) -> Tuple[torch.Tensor, torch.Tensor, torch.Tensor]:
-        batch_size = X.size(0)
-
-        X_imputed = self.implicit_imputation_layer(X, missing_mask)
-
-        hidden_state = torch.zeros(
-            (batch_size, self.d_rnn_hidden), dtype=X.dtype, device=X.device
-        )
-        cell_state = torch.zeros(
-            (batch_size, self.d_rnn_hidden), dtype=X.dtype, device=X.device
-        )
-        # cell_state_collector = torch.empty((batch_size, self.n_steps, self.d_rnn_hidden),
-        #                                    dtype=X.dtype, device=X.device)
-        for i in range(self.n_steps):
-            x = X_imputed[:, i, :]
-            hidden_state, cell_state = self.encoder(x, (hidden_state, cell_state))
-            # cell_state_collector[:, i, :] = cell_state
-
-        cell_state_collector = self.ae_encode_layers(cell_state)
-        mu_tilde = self.mu_layer(cell_state_collector)
-        stddev_tilde = self.stddev_layer(cell_state_collector)
-        z = self.z_sampling(mu_tilde, stddev_tilde)
-        return z, mu_tilde, stddev_tilde
-
-    def decode(self, z: torch.Tensor) -> torch.Tensor:
-        hidden_state = z
-        hidden_state = self.ae_decode_layers(hidden_state)
-
-        cell_state = torch.zeros(hidden_state.size(), dtype=z.dtype, device=z.device)
-        inputs = torch.zeros(
-            (z.size(0), self.n_steps, self.d_input), dtype=z.dtype, device=z.device
-        )
-
-        hidden_state_collector = torch.empty(
-            (z.size(0), self.n_steps, self.d_rnn_hidden), dtype=z.dtype, device=z.device
-        )
-        for i in range(self.n_steps):
-            x = inputs[:, i, :]
-            hidden_state, cell_state = self.decoder(x, (hidden_state, cell_state))
-            hidden_state_collector[:, i, :] = hidden_state
-
-        reconstruction = self.rnn_transform_layer(hidden_state_collector)
-        return reconstruction
-
-    def get_results(
-        self, X: torch.Tensor, missing_mask: torch.Tensor
-    ) -> Tuple[
-        torch.Tensor,
-        torch.Tensor,
-        torch.Tensor,
-        torch.Tensor,
-        torch.Tensor,
-        torch.Tensor,
-        torch.Tensor,
-    ]:
-        z, mu_tilde, stddev_tilde = self.encode(X, missing_mask)
-        X_reconstructed = self.decode(z)
-        mu_c, var_c, phi_c = self.gmm_layer()
-        return X_reconstructed, mu_c, var_c, phi_c, z, mu_tilde, stddev_tilde
-
-    def forward(
-        self,
-        inputs: dict,
-        pretrain: bool = False,
-        training: bool = True,
-    ) -> dict:
-        X, missing_mask = inputs["X"], inputs["missing_mask"]
-        device = X.device
-
-        (
-            X_reconstructed,
-            mu_c,
-            var_c,
-            phi_c,
-            z,
-            mu_tilde,
-            stddev_tilde,
-        ) = self.get_results(X, missing_mask)
-        imputed_X = X_reconstructed * (1 - missing_mask) + X * missing_mask
-
-        if not training and not pretrain:
-            results = {
-                "mu_tilde": mu_tilde,
-                "stddev_tilde": stddev_tilde,
-                "mu": mu_c,
-                "var": var_c,
-                "phi": phi_c,
-                "z": z,
-                "imputed_X": imputed_X,
-            }
-            # if only run clustering, then no need to calculate loss
-            return results
-
-        # calculate the reconstruction loss
-        unscaled_reconstruction_loss = cal_mse(X_reconstructed, X, missing_mask)
-        reconstruction_loss = (
-            unscaled_reconstruction_loss
-            * self.n_steps
-            * self.d_input
-            / missing_mask.sum()
-        )
-        if pretrain:
-            results = {"loss": reconstruction_loss, "z": z}
-            return results
-
-        # calculate the latent loss
-        var_tilde = torch.exp(stddev_tilde)
-        stddev_c = torch.log(var_c + self.eps)
-        log_2pi = torch.log(torch.tensor([2 * torch.pi], device=device))
-        log_phi_c = torch.log(phi_c + self.eps)
-
-        batch_size = z.shape[0]
-
-        ii, jj = torch.meshgrid(
-            torch.arange(self.n_clusters, dtype=torch.int64, device=device),
-            torch.arange(batch_size, dtype=torch.int64, device=device),
-            indexing="ij",
-        )
-        ii = ii.flatten()
-        jj = jj.flatten()
-
-        lsc_b = stddev_c.index_select(dim=0, index=ii)
-        mc_b = mu_c.index_select(dim=0, index=ii)
-        sc_b = var_c.index_select(dim=0, index=ii)
-        z_b = z.index_select(dim=0, index=jj)
-        log_pdf_z = -0.5 * (lsc_b + log_2pi + torch.square(z_b - mc_b) / sc_b)
-        log_pdf_z = log_pdf_z.reshape([batch_size, self.n_clusters, self.d_mu_stddev])
-
-        log_p = log_phi_c + log_pdf_z.sum(dim=2)
-        lse_p = log_p.logsumexp(dim=1, keepdim=True)
-        log_gamma_c = log_p - lse_p
-        gamma_c = torch.exp(log_gamma_c)
-
-        term1 = torch.log(var_c + self.eps)
-        st_b = var_tilde.index_select(dim=0, index=jj)
-        sc_b = var_c.index_select(dim=0, index=ii)
-        term2 = torch.reshape(
-            st_b / (sc_b + self.eps), [batch_size, self.n_clusters, self.d_mu_stddev]
-        )
-        mt_b = mu_tilde.index_select(dim=0, index=jj)
-        mc_b = mu_c.index_select(dim=0, index=ii)
-        term3 = torch.reshape(
-            torch.square(mt_b - mc_b) / (sc_b + self.eps),
-            [batch_size, self.n_clusters, self.d_mu_stddev],
-        )
-
-        latent_loss1 = 0.5 * torch.sum(
-            gamma_c * torch.sum(term1 + term2 + term3, dim=2), dim=1
-        )
-        latent_loss2 = -torch.sum(gamma_c * (log_phi_c - log_gamma_c), dim=1)
-        latent_loss3 = -0.5 * torch.sum(1 + stddev_tilde, dim=1)
-
-        latent_loss1 = latent_loss1.mean()
-        latent_loss2 = latent_loss2.mean()
-        latent_loss3 = latent_loss3.mean()
-        latent_loss = latent_loss1 + latent_loss2 + latent_loss3
-
-        results = {
-            "loss": reconstruction_loss + self.alpha * latent_loss,
-            "z": z,
-            "imputed_X": imputed_X,
-        }
-
-        return results
 
 
 class VaDER(BaseNNClusterer):
@@ -624,7 +381,7 @@ class VaDER(BaseNNClusterer):
         var_collector = []
         phi_collector = []
         z_collector = []
-        imputed_X_collector = []
+        imputation_latent_collector = []
         clustering_results_collector = []
 
         with torch.no_grad():
@@ -644,8 +401,8 @@ class VaDER(BaseNNClusterer):
                 phi_collector.append(phi)
                 z = results["z"].cpu().numpy()
                 z_collector.append(z)
-                imputed_X = results["imputed_X"].cpu().numpy()
-                imputed_X_collector.append(imputed_X)
+                imputation_latent = results["imputation_latent"].cpu().numpy()
+                imputation_latent_collector.append(imputation_latent)
 
                 def func_to_apply(
                     mu_t_: np.ndarray,
@@ -676,7 +433,7 @@ class VaDER(BaseNNClusterer):
             "var": np.concatenate(var_collector),
             "phi": np.concatenate(phi_collector),
             "z": np.concatenate(z_collector),
-            "imputation": np.concatenate(imputed_X_collector),
+            "imputation_latent": np.concatenate(imputation_latent_collector),
         }
 
         result_dict = {
